ci(danger): add check ensuring gravitee node version is consistent
